make StageDateSuggestionResolver dependency-graph driven; remove chronology inference from display order; make BM derive its boundary from BID, not TEC; make COB use the later completion across TEC and BM; route PO projected-state chronology through the common resolver; route HoD approval chronology through the same resolver; make timeline/actuals/inferred-start paths use the same graph semantics; align V1/V2 fallback dependencies with StageFlowSeeder; remove the unsafe ordered-stage resolver overload; preserve same-day commencement and next-day suggestions.
Resulting logic

The lifecycle is now interpreted as:

                 ┌── TEC ──┐
BID ─────────────┤         ├── COB
                 └── BM ───┘

Therefore, with your actual example:

BID completed        06 Aug 2026

TEC
Start                07 Aug
Complete             21 Aug

BM
Start                07/10 Aug   ← VALID
Complete             21 Aug etc.

PRISM will not use TEC completion on 21 Aug as BM's start boundary. BM derives its chronology from BID:

Earliest BM start    06 Aug
Suggested BM start   07 Aug

For COB, both parallel branches are respected. If TEC completes 21 Aug and BM completes 25 Aug:

Earliest COB start   25 Aug
Suggested COB start  26 Aug

+++ ProjectManagement.Tests/Publications/CompendiumPhase41ProductionConvergenceTests.cs
@@ -13,7 +13,10 @@ public sealed class CompendiumPhase41ProductionConvergenceTests
             * CompendiumLayoutMetrics.ProjectBodyMaximumNarrativeScale
             * CompendiumLayoutMetrics.ProjectBodyLineHeightMultiplier;
 
-        Assert.Equal(expectedMaximumLine, CompendiumLayoutMetrics.MaximumProjectBodyLineHeightPoints, 3);
+        Assert.Equal(
+            (double)expectedMaximumLine,
+            (double)CompendiumLayoutMetrics.MaximumProjectBodyLineHeightPoints,
+            3);
         Assert.True(CompendiumLayoutMetrics.PhysicalPaginationNativeShapingTolerancePoints >= 2f);
         Assert.True(
             CompendiumLayoutMetrics.PhysicalPaginationReservePoints
